+++ android/Staccato_AN/app/src/main/java/com/on/staccato/data/network/TokenManager.kt
@@ -1,4 +1,4 @@
-package com.on.staccato.data
+package com.on.staccato.data.network
 
 import com.on.staccato.StaccatoApplication
 import kotlinx.coroutines.Dispatchers
